Day 17, part 1
This took _forever_ to finish. Calculating the paths was trivial,
even taking into account the fact that I learned about generateSequence()
to make it a functional-style solution.

Figuring out how to limit the search space was the hard part. Limiting
the dx values was the first breakthrough, after far too much time.
Limiting the dy values was harder, because I took myself down a dead
end, because in the sample sets, any dx that didn't hit at dy = 0
wouldn't hit at all. In the full problem set, that's not true.