Day 22
This one is almost not my solution. I thought it through as far as
"keep a list of disjoint regions where everything is on" but couldn't
wrap my head around how to split regions up. So, I went to the
Reddit thread for hints, and ultimately wound up adapting a Python
solution.