Day 24 ... sort of
I wrote a brute-force solver for part one, and unsurprisingly,
that won't do the trick. The way to solve this is to analyze the
MONAD program, figure out what it's doing, and use that knowledge
to find a better solution.

But frankly, that's not how I want to spend my day.